% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmr2.R
\name{xmr2}
\alias{xmr2}
\title{Tidyeval Version of xmr()}
\usage{
xmr2(dataframe, measure, ...)
}
\arguments{
\item{dataframe}{The dataframe or tibble to calculate from.
Data must be in a tidy format.
At least one variable for time and one variable for measure.}

\item{measure}{The column containing the measure. Must be in numeric format.}

\item{...}{Arguments to pipe to xmr}
}
\description{
Used to calculate XMR data. Now works with more tidy workflows.
}
