% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmr_chart2.R
\name{xmr_chart2}
\alias{xmr_chart2}
\title{Generate the XMR chart for XMR data.}
\usage{
xmr_chart2(dataframe, time, measure, ...)
}
\arguments{
\item{dataframe}{Output from xmR()}

\item{time}{Time column}

\item{measure}{Measure}

\item{...}{Arguments to pipe to xmr_chart()}
}
\description{
Useful for diagnostics on xmr, and just visualizing the data. Now works with more tidy workflows.
}
